+++ docxbox/app/app_log.h
@@ -4,6 +4,7 @@
 #ifndef DOCXBOX_APP_APP_LOG_H_
 #define DOCXBOX_APP_APP_LOG_H_
 
+#include <docxbox/helper/helper_dateTime.h>
 #include <docxbox/helper/helper_file.h>
 
 #include <iostream>
@@ -11,6 +12,8 @@
 #include <vector>
 
 namespace docxbox {
+
+static const char *const kFormatDateTimeLog = "%Y-%m-%d %H:%M:%S";
 
 class AppLog {
  public:
@@ -29,33 +32,46 @@ class AppLog {
 
   static void DeleteInstance();
 
-  static bool Error(const std::string &message);
-  static bool Info(const std::string &message);
+  // Store given message for later output to stdout / logout
+  void Notify(const std::string& message, const std::string& type);
+
+  static bool NotifyError(const std::string &message);
+  static bool NotifyInfo(const std::string &message, bool file_only = false);
+
+  // Log docxBox execution arguments to log file
+  static void LogStartUp(int argc, char *const *argv);
 
   static void Output(bool delete_instance = true);
+  void OutputToStdOut();
+  void OutputToLogFile();
+
+  static bool IsSilent();
 
  private:
   // Singleton instance pointer
   static AppLog* m_pThis_;
 
   std::vector<std::string> messages_;
-  // TODO(kay): collect dateTimes of notifications
-  //            and prefix messages in out.log with them
   std::vector<std::string> timestamps_;
+
+  // Messages that are not to be output to stdout, but only to log file
+  std::vector<bool> file_only_messages_;
 
   LogMode mode_ = LogTo_StdOut;
 
   std::string path_log_file_;
 
- private:
   // Copy constructor
-  AppLog(const AppLog&){};
+  AppLog(const AppLog&) {}
 
   // Assignment operator
-  AppLog& operator=(const AppLog&){ return *this; };
+  AppLog& operator = (const AppLog&) { return *this; }
 
   void InitMode();
   void InitLogFile();
+
+  // Remember current dateTime (notification occurred)
+  void PushBackTime();
 };
 
 }  // namespace docxbox