+++ docxbox/helper/helper_dateTime.cc
@@ -13,7 +13,7 @@ std::string DateTime::GetCurrentDateTimeFormatted(const char *format) {
   char buffer[80];
 
   time(&raw_time);
-  time_info = localtime(&raw_time);
+  time_info = localtime(&raw_time);  // NOLINT [build/c++11]
 
   strftime(buffer, sizeof(buffer), format, time_info);
   std::string str(buffer);
@@ -24,7 +24,7 @@ std::string DateTime::GetCurrentDateTimeFormatted(const char *format) {
 // @return dateTime in ISO 8601 format, e.g. "1994-11-05T13:15:30Z"
 // @see ISO-8601 https://www.w3.org/TR/NOTE-datetime
 std::string DateTime::GetCurrentDateTimeInIso8601() {
-  return GetCurrentDateTimeFormatted(kFormatIso8601DateTime);
+  return GetCurrentDateTimeFormatted(kFormatDateTimeIso8601);
 }
 
 // Verify given string being ISO8601 date. Ex: 2016-02-22T10:31:00Z