+++ docxbox/helper/helper_xml.cc
@@ -15,7 +15,7 @@ bool Xml::IndentXmlFile(const std::string& path_xml) {
   if (doc.ErrorID() != 0) return false;
 
   return tinyxml2::XML_SUCCESS != doc.SaveFile(path_xml.c_str())
-         ? docxbox::AppLog::Error("Failed saving: " + path_xml)
+         ? docxbox::AppLog::NotifyError("Failed saving: " + path_xml)
          : true;
 }